+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next
 
-- Add a zoomToGene API that allows to zoom a HiGlass view to a location near a certain gene.
+- Add a `zoomToGene` API that allows to zoom a HiGlass view to a location near a certain gene.
+- Add a `suggestGene` API that returns a list of genes of top match for a given keyword based on `autocompleteServer`.
 
 _[Detailed changes since v1.11.4](https://github.com/higlass/higlass/compare/v1.11.4...develop)_